merge

+++ src/main/java/site/alice/liveman/job/RecordAndCleanUpJob.java
@@ -20,6 +20,7 @@ package site.alice.liveman.job;
 
 import de.tuberlin.onedrivesdk.folder.OneFolder;
 import de.tuberlin.onedrivesdk.uploadFile.OneUploadFile;
+import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang.StringUtils;
@@ -38,7 +39,6 @@ import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.OutputStream;
 import java.text.SimpleDateFormat;
-import java.util.Arrays;
 import java.util.Comparator;
 import java.util.LinkedList;
 import java.util.List;
@@ -50,6 +50,7 @@ public class RecordAndCleanUpJob {
 
     private static final Logger              LOGGER                    = LoggerFactory.getLogger(RecordAndCleanUpJob.class);
     private static final long                LAST_MODIFIED_TIME_MILLIS = 15 * 60 * 1000;
+    private static final long                MAX_RECORD_FILE_SIZE      = 4 * 1024 * 1024 * 1024L;
     @Autowired
     private              MediaHistoryService mediaHistoryService;
     @Autowired
@@ -57,7 +58,7 @@ public class RecordAndCleanUpJob {
     @Autowired
     private              OneDriveUtil        oneDriveUtil;
 
-    @Scheduled(cron = "0 0/1 * * * ?")
+    @Scheduled(cron = "0 0/5 * * * ?")
     public void recordAndCleanUpJob() {
         LOGGER.info("开始查找需要上传录像和清理的媒体文件...");
         if (StringUtils.isEmpty(liveManSetting.getOneDriveToken())) {
@@ -70,7 +71,7 @@ public class RecordAndCleanUpJob {
             for (File mediaDir : listDir) {
                 File[] pathList = mediaDir.listFiles(file -> System.currentTimeMillis() - file.lastModified() >= LAST_MODIFIED_TIME_MILLIS);
                 if (pathList != null) {
-                    Arrays.stream(pathList).parallel().forEach(sourcePath -> {
+                    for (File sourcePath : pathList) {
                         LOGGER.info("开始上传[videoId=" + sourcePath.getName() + "]的录像数据...");
                         switch (mediaDir.getName()) {
                             case "m3u8": {
@@ -83,12 +84,26 @@ public class RecordAndCleanUpJob {
                                 uploadDir(sourcePath, sourcePath.getName());
                                 break;
                             }
+                            case "broadcast": {
+                                FileUtils.deleteQuietly(sourcePath);
+                                break;
+                            }
                         }
-                    });
+                    }
                 }
             }
         }
         LOGGER.info("上传录像和清理任务已完成！");
+    }
+
+    @Scheduled(cron = "0 0/1 * * * ?")
+    public void freeSpaceMasterJob() {
+        String mediaTempPath = liveManSetting.getTempPath();
+        File mediaTempDir = new File(mediaTempPath);
+        if (mediaTempDir.getFreeSpace() < 1073741824) {
+            LOGGER.warn("磁盘可用空间严重不足，强制清除所有录像缓存数据！当前可用空间:" + mediaTempDir.getFreeSpace() + "字节");
+            FileUtils.deleteQuietly(mediaTempDir);
+        }
     }
 
     private void uploadDir(File sourceFile, final String videoId) {
@@ -109,35 +124,39 @@ public class RecordAndCleanUpJob {
                 LOGGER.info("文件[" + sourceFile + "]的长度为0，跳过上传!");
                 success = true;
             } else if (!StringUtils.isEmpty(liveManSetting.getOneDriveToken())) {
-                for (int i = 0; i < 3; i++) {
-                    try {
-                        OneFolder recordFolder = oneDriveUtil.getOneFolder("Record");
-                        SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
-                        SimpleDateFormat timeFormat = new SimpleDateFormat("HH_mm_ss");
-                        String fileName = mediaHistory.getChannelName().trim();
-                        if (StringUtils.isEmpty(fileName)) {
-                            fileName = "手动推流";
-                        }
-                        fileName += "\\/" + dateFormat.format(mediaHistory.getDatetime());
-                        fileName += "\\/" + timeFormat.format(mediaHistory.getDatetime()) + "_" + mediaHistory.getVideoId() + "_" + replaceFileName(String.valueOf(mediaHistory.getVideoTitle()).trim());
-                        fileName += "\\/" + sourceFile.getName();
-                        LOGGER.info("开始上传录像[" + sourceFile.getAbsoluteFile() + "]到[" + fileName + "]...");
-                        OneUploadFile oneUploadFile = recordFolder.uploadFile(sourceFile, fileName);
-                        oneUploadFile.startUpload();
-                        LOGGER.info("录像文件[" + sourceFile.getAbsoluteFile() + "]上传已完成！");
-                        success = true;
-                        break;
-                    } catch (Throwable throwable) {
-                        LOGGER.error("录像保存失败[" + sourceFile.getAbsoluteFile() + "]，重试(" + (i + 1) + "/3)", throwable);
+                try {
+                    OneFolder recordFolder = oneDriveUtil.getOneFolder("Record");
+                    SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
+                    SimpleDateFormat timeFormat = new SimpleDateFormat("HH_mm_ss");
+                    String fileName = mediaHistory.getChannelName();
+                    if (StringUtils.isEmpty(fileName) || fileName.equals("null")) {
+                        fileName = "手动推流";
                     }
+                    fileName = fileName.trim();
+                    fileName += "\\/" + dateFormat.format(mediaHistory.getDatetime());
+                    fileName += "\\/" + timeFormat.format(mediaHistory.getDatetime()) + "_" + mediaHistory.getVideoId() + "_" + replaceFileName(String.valueOf(mediaHistory.getVideoTitle()).trim());
+                    fileName += "\\/" + sourceFile.getName();
+                    LOGGER.info("开始上传录像[" + sourceFile.getAbsoluteFile() + "]到[" + fileName + "]...");
+                    OneUploadFile oneUploadFile = recordFolder.uploadFile(sourceFile, fileName);
+                    oneUploadFile.startUpload();
+                    LOGGER.info("录像文件[" + sourceFile.getAbsoluteFile() + "]上传已完成！");
+                    success = true;
+                } catch (Throwable throwable) {
+                    LOGGER.error("录像保存失败[" + sourceFile.getAbsoluteFile() + "]", throwable);
                 }
+            } else {
+                success = true;
             }
         }
-        if (success) {
-            sourceFile.delete();
+        if (sourceFile.exists()) {
+            if (success) {
+                sourceFile.delete();
+            } else if (sourceFile.getFreeSpace() < MAX_RECORD_FILE_SIZE) {
+                LOGGER.warn("磁盘可用空间不足4GB，当前可用空间[" + sourceFile.getFreeSpace() + "]，强制删除上传失败的录像文件[" + sourceFile + "]");
+                sourceFile.delete();
+            }
         }
     }
-
 
     private String replaceFileName(String fileName) {
         Pattern pattern = Pattern.compile("[#\\\\/:\\*\\?\\\"<>\\|]");
@@ -158,12 +177,26 @@ public class RecordAndCleanUpJob {
                         LOGGER.info(fileName + " is not a seq file skipped.");
                     }
                 }
-                seqList.sort(Comparator.naturalOrder());
-                try (OutputStream os = new FileOutputStream(m3u8Path + "/index.ts", true)) {
+                if (!seqList.isEmpty()) {
+                    seqList.sort(Comparator.naturalOrder());
+                    int i = 0;
+                    long totalSize = 0;
+                    OutputStream os = null;
                     for (Integer integer : seqList) {
                         File tsFile = new File(m3u8Path + "/" + integer + ".ts");
                         FileInputStream is = new FileInputStream(tsFile);
-                        IOUtils.copy(is, os);
+                        if (os == null || totalSize >= MAX_RECORD_FILE_SIZE) {
+                            File outputFile = null;
+                            while (outputFile == null || (outputFile.exists() && outputFile.length() >= MAX_RECORD_FILE_SIZE)) {
+                                outputFile = new File(m3u8Path + "/index_" + i++ + ".ts");
+                            }
+                            if (os != null) {
+                                os.close();
+                            }
+                            os = new FileOutputStream(outputFile, true);
+                            totalSize = outputFile.length();
+                        }
+                        totalSize += IOUtils.copyLarge(is, os);
                         is.close();
                         tsFile.delete();
                     }
